requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt[crypto]>=5.3.1
drf-spectacular>=0.27.0
django-phonenumber-field[phonenumberslite]>=8.0.0
django-cors-headers>=4.0
dj-database-url>=2.1
python-dotenv>=1.0
